+++ CHANGELOG.md
@@ -158,3 +158,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Removed
 - Delete the implementations for the following traits
     - Boffow<[u8]> for IdRef
+    - DerefMut for IdRef